refactor: extract FiltersStock
